Move motion blinds coordinator to its own file
